Rename "display.manifests" to "display.bags"
