+++ display/src/main/scala/uk/ac/wellcome/platform/archive/display/bags/DisplayStorageManifest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.display.manifests
+package uk.ac.wellcome.platform.archive.display.bags
 
 import java.net.URL
